REFACTOR: rename DayNightCycleManager to DayNightPluginManager
